Notebook and Report for model selection
Created the notebook and ran it multiple times to arrive at the best run for different models. Then created a report with the results and analysis of best model. We will use a separate notebook to further create tables using the chosen model for the final report. Overfitting was explored for multiple models and documented in the report.